+++ src/func/func_serialize.c
@@ -9,7 +9,7 @@ int mli_Func_to_io(const struct mli_Func *func, struct mli_IO *f)
         chk(mli_MagicId_set(&magic, "mli_Func"));
         chk_IO_write(&magic, sizeof(struct mli_MagicId), 1u, f);
 
-        chk_IO_write(&func->num_points, sizeof(uint32_t), 1u, f);
+        chk_IO_write(&func->num_points, sizeof(uint64_t), 1u, f);
         chk_IO_write(func->x, sizeof(double), func->num_points, f);
         chk_IO_write(func->y, sizeof(double), func->num_points, f);
         return 1;
@@ -19,13 +19,13 @@ chk_error:
 
 int mli_Func_from_io(struct mli_Func *func, struct mli_IO *f)
 {
-        uint32_t num_points;
+        uint64_t num_points;
         struct mli_MagicId magic;
         chk_IO_read(&magic, sizeof(struct mli_MagicId), 1u, f);
         chk(mli_MagicId_has_word(&magic, "mli_Func"));
         mli_MagicId_warn_version(&magic);
 
-        chk_IO_read(&num_points, sizeof(uint32_t), 1u, f);
+        chk_IO_read(&num_points, sizeof(uint64_t), 1u, f);
         chk(mli_Func_malloc(func, num_points));
         chk_IO_read(func->x, sizeof(double), func->num_points, f);
         chk_IO_read(func->y, sizeof(double), func->num_points, f);
